refactor code to different files
