+++ src/dom.js
@@ -1,0 +1,16 @@
+const projectName = document.querySelector('#name');
+const projectForm = document.querySelector('#project-form');
+const projectsDiv = document.querySelector('.projects');
+const dropdown = document.querySelector('#select');
+const toDoForm = document.querySelector('#todo-form');
+
+const title = document.querySelector('#title');
+const description = document.querySelector('#description');
+const select = document.querySelector('#select');
+const dueDate = document.querySelector('#due-date');
+const priority = document.querySelector('#priority');
+
+export {
+  projectName, projectForm, projectsDiv, dropdown, toDoForm, title,
+  description, select, dueDate, priority,
+};
